chore: clear main.go
