+++ handlers/uiapi.go
@@ -1,0 +1,368 @@
+package handlers
+
+import (
+	"fmt"
+	"log"
+	"net/http"
+	"sort"
+	"strconv"
+	"strings"
+	"time"
+
+	"github.com/dadosjusbr/api/models"
+	"github.com/dadosjusbr/api/services"
+	"github.com/dadosjusbr/storage"
+	strModels "github.com/dadosjusbr/storage/models"
+	"github.com/gocarina/gocsv"
+	"github.com/labstack/echo"
+)
+
+type UiApiHandler struct {
+	Client           storage.Client
+	PostgresDb       services.PostgresDB
+	Sess             *services.AwsSession
+	S3Bucket         string
+	Loc              *time.Location
+	EnvOmittedFields []string
+	SearchLimit      int
+	DownloadLimit    int
+}
+
+func (a UiApiHandler) GetSummaryOfAgency(c echo.Context) error {
+	year, err := strconv.Atoi(c.Param("ano"))
+	if err != nil {
+		return c.JSON(http.StatusBadRequest, fmt.Sprintf("Parâmetro ano=%d inválido", year))
+	}
+	month, err := strconv.Atoi(c.Param("mes"))
+	if err != nil {
+		return c.JSON(http.StatusBadRequest, fmt.Sprintf("Parâmetro mês=%d", month))
+	}
+	agencyName := c.Param("orgao")
+	agencyMonthlyInfo, agency, err := a.Client.GetOMA(month, year, agencyName)
+	if err != nil {
+		return c.JSON(http.StatusBadRequest, fmt.Sprintf("Parâmetro ano=%d, mês=%d ou nome do orgão=%s são inválidos", year, month, agencyName))
+	}
+	agencySummary := models.AgencySummary{
+		FullName:   agency.Name,
+		TotalWage:  agencyMonthlyInfo.Summary.BaseRemuneration.Total,
+		MaxWage:    agencyMonthlyInfo.Summary.BaseRemuneration.Max,
+		TotalPerks: agencyMonthlyInfo.Summary.OtherRemunerations.Total,
+		MaxPerk:    agencyMonthlyInfo.Summary.OtherRemunerations.Max,
+		TotalRemuneration: agencyMonthlyInfo.Summary.BaseRemuneration.Total +
+			agencyMonthlyInfo.Summary.OtherRemunerations.Total,
+		TotalMembers: agencyMonthlyInfo.Summary.Count,
+		CrawlingTime: agencyMonthlyInfo.CrawlingTimestamp,
+		HasNext:      time.Date(year, time.Month(month), 1, 0, 0, 0, 0, time.UTC).In(a.Loc).Before(time.Now().AddDate(0, 1, 0)),
+		HasPrevious:  time.Date(year, time.Month(month), 1, 0, 0, 0, 0, time.UTC).In(a.Loc).After(time.Date(2018, 1, 1, 0, 0, 0, 0, time.UTC).In(a.Loc)),
+	}
+	return c.JSON(http.StatusOK, agencySummary)
+}
+
+func (a UiApiHandler) GetSalaryOfAgencyMonthYear(c echo.Context) error {
+	month, err := strconv.Atoi(c.Param("mes"))
+	if err != nil {
+		return c.JSON(http.StatusBadRequest, fmt.Sprintf("Parâmetro mês=%d", month))
+	}
+	year, err := strconv.Atoi(c.Param("ano"))
+	if err != nil {
+		return c.JSON(http.StatusBadRequest, fmt.Sprintf("Parâmetro ano=%d", year))
+	}
+	agencyName := strings.ToLower(c.Param("orgao"))
+	agencyMonthlyInfo, _, err := a.Client.GetOMA(month, year, agencyName)
+	if err != nil {
+		log.Printf("[salary agency month year] error getting data for second screen(mes:%d ano:%d, orgao:%s):%q", month, year, agencyName, err)
+		return c.JSON(http.StatusBadRequest, fmt.Sprintf("Parâmetro ano=%d, mês=%d ou nome do orgão=%s são inválidos", year, month, agencyName))
+	}
+	if agencyMonthlyInfo.ProcInfo.String() != "" {
+		var newEnv = agencyMonthlyInfo.ProcInfo.Env
+		for _, omittedField := range a.EnvOmittedFields {
+			for i, field := range newEnv {
+				if strings.Contains(field, omittedField) {
+					newEnv[i] = omittedField + "= ##omitida##"
+					break
+				}
+			}
+		}
+		agencyMonthlyInfo.ProcInfo.Env = newEnv
+		return c.JSON(http.StatusPartialContent, models.ProcInfoResult{
+			ProcInfo:          agencyMonthlyInfo.ProcInfo,
+			CrawlingTimestamp: agencyMonthlyInfo.CrawlingTimestamp,
+		})
+	}
+	return c.JSON(http.StatusOK, models.DataForChartAtAgencyScreen{
+		Members:     agencyMonthlyInfo.Summary.IncomeHistogram,
+		MaxSalary:   agencyMonthlyInfo.Summary.BaseRemuneration.Max,
+		PackageURL:  agencyMonthlyInfo.Package.URL,
+		PackageHash: agencyMonthlyInfo.Package.Hash,
+		PackageSize: agencyMonthlyInfo.Package.Size,
+	})
+}
+
+func (a UiApiHandler) GetTotalsOfAgencyYear(c echo.Context) error {
+	year, err := strconv.Atoi(c.Param("ano"))
+	if err != nil {
+		return c.JSON(http.StatusBadRequest, fmt.Sprintf("Parâmetro ano=%d inválido", year))
+	}
+	aID := c.Param("orgao")
+	agenciesMonthlyInfo, err := a.Client.Db.GetMonthlyInfo([]strModels.Agency{{ID: aID}}, year)
+	if err != nil {
+		log.Printf("[totals of agency year] error getting data for first screen(ano:%d, estado:%s):%q", year, aID, err)
+		return c.JSON(http.StatusBadRequest, fmt.Sprintf("Parâmetro ano=%d ou orgao=%s inválidos", year, aID))
+	}
+	var monthTotalsOfYear []models.MonthTotals
+	agency, err := a.Client.Db.GetAgency(aID)
+	if err != nil {
+		log.Printf("[totals of agency year] error getting data for first screen(estado:%s):%q", aID, err)
+		return c.JSON(http.StatusBadRequest, fmt.Sprintf("Parâmetro orgao=%s inválido", aID))
+	}
+	host := c.Request().Host
+	agency.URL = fmt.Sprintf("%s/v1/orgao/%s", host, agency.ID)
+	for _, agencyMonthlyInfo := range agenciesMonthlyInfo[aID] {
+		if agencyMonthlyInfo.Summary != nil && agencyMonthlyInfo.Summary.BaseRemuneration.Total+agencyMonthlyInfo.Summary.OtherRemunerations.Total > 0 {
+			monthTotals := models.MonthTotals{Month: agencyMonthlyInfo.Month,
+				BaseRemuneration:   agencyMonthlyInfo.Summary.BaseRemuneration.Total,
+				OtherRemunerations: agencyMonthlyInfo.Summary.OtherRemunerations.Total,
+				CrawlingTimestamp:  agencyMonthlyInfo.CrawlingTimestamp,
+				TotalMembers:       agencyMonthlyInfo.Summary.Count,
+			}
+			monthTotalsOfYear = append(monthTotalsOfYear, monthTotals)
+
+			// The status 4 is a report from crawlers that data is unavailable or malformed. By removing them from the API results, we make sure they are displayed as if there is no data.
+		} else if agencyMonthlyInfo.ProcInfo.String() != "" && agencyMonthlyInfo.ProcInfo.Status != 4 {
+			monthTotals := models.MonthTotals{Month: agencyMonthlyInfo.Month,
+				BaseRemuneration:   0,
+				OtherRemunerations: 0,
+				CrawlingTimestamp:  agencyMonthlyInfo.CrawlingTimestamp,
+				Error:              &models.ProcError{Stdout: agencyMonthlyInfo.ProcInfo.Stdout, Stderr: agencyMonthlyInfo.ProcInfo.Stderr},
+			}
+			monthTotalsOfYear = append(monthTotalsOfYear, monthTotals)
+		}
+	}
+	sort.Slice(monthTotalsOfYear, func(i, j int) bool {
+		return monthTotalsOfYear[i].Month < monthTotalsOfYear[j].Month
+	})
+	destKey := fmt.Sprintf("%s/datapackage/%s-%d.zip", aID, aID, year)
+	bkp, _ := a.Client.Cloud.GetFile(destKey)
+	var pkg *strModels.Package
+	if bkp != nil {
+		pkg = &strModels.Package{
+			AgencyID: &aID,
+			Year:     &year,
+			Package:  *bkp,
+		}
+	}
+
+	agencyTotalsYear := models.AgencyTotalsYear{Year: year, Agency: agency, MonthTotals: monthTotalsOfYear, AgencyFullName: agency.Name, SummaryPackage: pkg}
+	return c.JSON(http.StatusOK, agencyTotalsYear)
+}
+
+func (a UiApiHandler) GetBasicInfoOfType(c echo.Context) error {
+	yearOfConsult := time.Now().Year()
+	groupName := c.Param("grupo")
+	var agencies []strModels.Agency
+	var err error
+	var estadual bool
+	var exists bool
+	jurisdicao := map[string]string{
+		"justica-eleitoral":    "Eleitoral",
+		"ministerios-publicos": "Ministério",
+		"justica-estadual":     "Estadual",
+		"justica-do-trabalho":  "Trabalho",
+		"justica-federal":      "Federal",
+		"justica-militar":      "Militar",
+		"justica-superior":     "Superior",
+		"conselhos-de-justica": "Conselho",
+	}
+
+	// Adaptando as URLs do site com o banco de dados
+	// Primeiro consultamos entre as chaves do mapa.
+	if jurisdicao[groupName] != "" {
+		groupName = jurisdicao[groupName]
+	} else {
+		// Caso não encontremos entre as chaves, verificamos entre os valores do mapa.
+		// Isso pois, até a consolidação ser finalizada, o front consulta a api com /Eleitoral, /Trabalho, etc.
+		for _, value := range jurisdicao {
+			if groupName == value {
+				exists = true
+				break
+			}
+		}
+		// Se a jurisdição não existir no mapa, verificamos se trata-se de um estado
+		if !exists {
+			values := map[string]struct{}{"AC": {}, "AL": {}, "AP": {}, "AM": {}, "BA": {}, "CE": {}, "DF": {}, "ES": {}, "GO": {}, "MA": {}, "MT": {}, "MS": {}, "MG": {}, "PA": {}, "PB": {}, "PR": {}, "PE": {}, "PI": {}, "RJ": {}, "RN": {}, "RS": {}, "RO": {}, "RR": {}, "SC": {}, "SP": {}, "SE": {}, "TO": {}}
+			if _, estadual = values[groupName]; estadual {
+				exists = true
+			}
+		}
+		// Se o parâmetro dado não for encontrado de forma alguma, retornamos um NOT FOUND (404)
+		if !exists {
+			return c.JSON(http.StatusNotFound, fmt.Sprintf("Grupo não encontrado: %s.", groupName))
+		}
+	}
+
+	if estadual {
+		agencies, err = a.Client.GetStateAgencies(groupName)
+	} else {
+		agencies, err = a.Client.GetOPJ(groupName)
+	}
+	if err != nil {
+		// That happens when there is no information on that year.
+		log.Printf("[basic info type] first error getting data for first screen(ano:%d, grupo:%s). Going to try again with last year:%q", yearOfConsult, groupName, err)
+		yearOfConsult = yearOfConsult - 1
+
+		if estadual {
+			agencies, err = a.Client.GetStateAgencies(groupName)
+		} else {
+			agencies, err = a.Client.GetOPJ(groupName)
+		}
+		if err != nil {
+			log.Printf("[basic info type] error getting data for first screen(ano:%d, grupo:%s):%q", yearOfConsult, groupName, err)
+			return c.JSON(http.StatusBadRequest, fmt.Sprintf("Parâmetros ano=%d ou grupo=%s são inválidos", yearOfConsult, groupName))
+		}
+	}
+	var agenciesBasic []models.AgencyBasic
+	for k := range agencies {
+		agenciesBasic = append(agenciesBasic, models.AgencyBasic{Name: agencies[k].ID, FullName: agencies[k].Name, AgencyCategory: agencies[k].Entity})
+	}
+	state := models.State{Name: c.Param("grupo"), ShortName: "", FlagURL: "", Agency: agenciesBasic}
+	return c.JSON(http.StatusOK, state)
+}
+
+func (a UiApiHandler) GetGeneralRemunerationFromYear(c echo.Context) error {
+	year, err := strconv.Atoi(c.Param("ano"))
+	if err != nil {
+		return c.JSON(http.StatusBadRequest, fmt.Sprintf("Parâmetro ano=%d inválido", year))
+	}
+	data, err := a.Client.Db.GetGeneralMonthlyInfosFromYear(year)
+	if err != nil {
+		fmt.Println("Error searching for monthly info from year: %w", err)
+		return c.JSON(http.StatusInternalServerError, fmt.Sprintf("Error buscando dados"))
+	}
+	return c.JSON(http.StatusOK, data)
+}
+
+func (a UiApiHandler) GeneralSummaryHandler(c echo.Context) error {
+	agencies, err := a.Client.GetAgenciesCount()
+	if err != nil {
+		return c.JSON(http.StatusInternalServerError, fmt.Sprintf("Erro ao contar orgãos: %q", err))
+	}
+	collections, err := a.Client.GetNumberOfMonthsCollected()
+	if err != nil {
+		return c.JSON(http.StatusInternalServerError, fmt.Sprintf("Erro ao contar registros: %q", err))
+	}
+	fmonth, fyear, err := a.Client.Db.GetFirstDateWithMonthlyInfo()
+	if err != nil {
+		log.Printf("Error buscando dados - GetFirstDateWithRemunerationRecords: %q", err)
+		return c.JSON(http.StatusInternalServerError, fmt.Sprintf("Erro buscando primeiro registro de remuneração: %q", err))
+	}
+	fdate := time.Date(fyear, time.Month(fmonth), 2, 0, 0, 0, 0, time.UTC).In(a.Loc)
+	lmonth, lyear, err := a.Client.GetLastDateWithMonthlyInfo()
+	if err != nil {
+		log.Printf("Error buscando dados - GetLastDateWithRemunerationRecords: %q", err)
+		return c.JSON(http.StatusInternalServerError, fmt.Sprintf("Erro buscando último registro de remuneração: %q", err))
+	}
+	ldate := time.Date(lyear, time.Month(lmonth), 2, 0, 0, 0, 0, time.UTC).In(a.Loc)
+	remuValue, err := a.Client.Db.GetGeneralMonthlyInfo()
+	if err != nil {
+		log.Printf("Error buscando dados - GetGeneralRemunerationValue: %q", err)
+		return c.JSON(http.StatusInternalServerError, fmt.Sprintf("Erro buscando valor total de remuneração: %q", err))
+	}
+	return c.JSON(http.StatusOK, models.GeneralTotals{
+		AgencyAmount:             int(agencies),
+		MonthlyTotalsAmount:      int(collections),
+		StartDate:                fdate,
+		EndDate:                  ldate,
+		RemunerationRecordsCount: int(collections),
+		GeneralRemunerationValue: remuValue,
+	})
+}
+
+func (a UiApiHandler) SearchByUrl(c echo.Context) error {
+	//Pegando os query params
+	years := c.QueryParam("anos")
+	months := c.QueryParam("meses")
+	agencies := c.QueryParam("orgaos")
+	categories := c.QueryParam("categorias")
+	types := c.QueryParam("tipos")
+	//Criando os filtros a partir dos query params e validando eles
+	filter, err := models.NewFilter(years, months, agencies, categories, types)
+	if err != nil {
+		return c.JSON(http.StatusBadRequest, err.Error())
+	}
+	var category string
+	if filter != nil {
+		category = filter.Category
+	}
+	// Pegando os resultados da pesquisa a partir dos filtros;
+	results, err := a.PostgresDb.Filter(a.PostgresDb.RemunerationQuery(filter), a.PostgresDb.Arguments(filter))
+	if err != nil {
+		log.Printf("Error querying BD (filter or counter):%q", err)
+		return c.JSON(http.StatusInternalServerError, err.Error())
+	}
+	remunerations, numRows, err := a.getSearchResults(a.SearchLimit, category, results)
+	if err != nil {
+		log.Printf("Error getting search results: %q", err)
+		return c.JSON(http.StatusInternalServerError, err.Error())
+	}
+
+	response := models.SearchResponse{
+		DownloadAvailable:  numRows > 0 && numRows <= a.DownloadLimit,
+		NumRowsIfAvailable: numRows,
+		DownloadLimit:      a.DownloadLimit,
+		SearchLimit:        a.SearchLimit,
+		Results:            remunerations, // retornando os SearchLimit primeiros elementos.
+	}
+	return c.JSON(http.StatusOK, response)
+}
+
+func (a UiApiHandler) DownloadByUrl(c echo.Context) error {
+	//Pegando os query params
+	years := c.QueryParam("anos")
+	months := c.QueryParam("meses")
+	agencies := c.QueryParam("orgaos")
+	categories := c.QueryParam("categorias")
+	types := c.QueryParam("tipos")
+
+	//Criando os filtros a partir dos query params e validando eles
+	filter, err := models.NewFilter(years, months, agencies, categories, types)
+	if err != nil {
+		return c.JSON(http.StatusBadRequest, err.Error())
+	}
+
+	results, err := a.PostgresDb.Filter(a.PostgresDb.RemunerationQuery(filter), a.PostgresDb.Arguments(filter))
+	if err != nil {
+		return c.JSON(http.StatusInternalServerError, err.Error())
+	}
+	searchResults, _, err := a.getSearchResults(a.DownloadLimit, filter.Category, results)
+	if err != nil {
+		return c.JSON(http.StatusInternalServerError, err.Error())
+	}
+
+	c.Response().Header().Set("Content-Disposition", "attachment; filename=dadosjusbr-remuneracoes.csv")
+	c.Response().Header().Set("Content-Type", c.Response().Header().Get("Content-Type"))
+	err = gocsv.Marshal(searchResults, c.Response().Writer)
+	if err != nil {
+		return c.JSON(http.StatusInternalServerError, fmt.Errorf("erro tentando fazer download do csv: %q", err))
+	}
+	return nil
+}
+
+func (a UiApiHandler) getSearchResults(limit int, category string, results []models.SearchDetails) ([]models.SearchResult, int, error) {
+	searchResults := []models.SearchResult{}
+	numRows := 0
+	if len(results) == 0 {
+		return searchResults, numRows, nil
+	} else {
+		// A razão para essa ordenação é que quando o usuário escolhe diversos órgãos
+		// provavelmente ele prefere ver dados de todos eles. Dessa forma, aumentamos
+		// as chances do preview limitado retornar dados de diversos órgãos.
+		sort.SliceStable(results, func(i, j int) bool {
+			return results[i].Ano < results[j].Ano || results[i].Mes < results[j].Mes
+		})
+		searchResults, numRows, err := a.Sess.GetRemunerationsFromS3(limit, a.DownloadLimit, category, a.S3Bucket, results)
+		if err != nil {
+			return nil, numRows, fmt.Errorf("failed to get remunerations from s3 %q", err)
+		}
+		return searchResults, numRows, nil
+	}
+}
