+++ main.go
@@ -5,17 +5,13 @@ import (
 	"log"
 	"net/http"
 	"os"
-	"sort"
-	"strconv"
-	"strings"
 	"time"
 
-	"github.com/dadosjusbr/api/models"
+	"github.com/dadosjusbr/api/handlers"
+	"github.com/dadosjusbr/api/services"
 	"github.com/dadosjusbr/storage"
-	strModels "github.com/dadosjusbr/storage/models"
 	"github.com/dadosjusbr/storage/repo/database"
 	"github.com/dadosjusbr/storage/repo/file_storage"
-	"github.com/gocarina/gocsv"
 	"github.com/joho/godotenv"
 	"github.com/newrelic/go-agent/v3/integrations/nrecho-v3"
 	"github.com/newrelic/go-agent/v3/newrelic"
@@ -67,8 +63,8 @@ type config struct {
 var pgS3Client *storage.Client
 var loc *time.Location
 var conf config
-var postgresDb *PostgresDB
-var sess *AwsSession
+var postgresDb *services.PostgresDB
+var sess *services.AwsSession
 
 // newClient takes a config struct and creates a client to connect with DB and Cloud5
 func newClient(db database.Interface, cloud file_storage.Interface) (*storage.Client, error) {
@@ -93,548 +89,6 @@ func newS3Client(c config) (*file_storage.S3Client, error) {
 		return nil, fmt.Errorf("error creating S3 client: %v", err.Error())
 	}
 	return s3Client, nil
-}
-
-func getTotalsOfAgencyYear(c echo.Context) error {
-	year, err := strconv.Atoi(c.Param("ano"))
-	if err != nil {
-		return c.JSON(http.StatusBadRequest, fmt.Sprintf("Parâmetro ano=%d inválido", year))
-	}
-	aID := c.Param("orgao")
-	agenciesMonthlyInfo, err := pgS3Client.Db.GetMonthlyInfo([]strModels.Agency{{ID: aID}}, year)
-	if err != nil {
-		log.Printf("[totals of agency year] error getting data for first screen(ano:%d, estado:%s):%q", year, aID, err)
-		return c.JSON(http.StatusBadRequest, fmt.Sprintf("Parâmetro ano=%d ou orgao=%s inválidos", year, aID))
-	}
-	var monthTotalsOfYear []models.MonthTotals
-	agency, err := pgS3Client.Db.GetAgency(aID)
-	if err != nil {
-		log.Printf("[totals of agency year] error getting data for first screen(estado:%s):%q", aID, err)
-		return c.JSON(http.StatusBadRequest, fmt.Sprintf("Parâmetro orgao=%s inválido", aID))
-	}
-	for _, agencyMonthlyInfo := range agenciesMonthlyInfo[aID] {
-		if agencyMonthlyInfo.Summary != nil && agencyMonthlyInfo.Summary.BaseRemuneration.Total+agencyMonthlyInfo.Summary.OtherRemunerations.Total > 0 {
-			monthTotals := models.MonthTotals{Month: agencyMonthlyInfo.Month,
-				BaseRemuneration:   agencyMonthlyInfo.Summary.BaseRemuneration.Total,
-				OtherRemunerations: agencyMonthlyInfo.Summary.OtherRemunerations.Total,
-				CrawlingTimestamp:  agencyMonthlyInfo.CrawlingTimestamp,
-				TotalMembers:       agencyMonthlyInfo.Summary.Count,
-			}
-			monthTotalsOfYear = append(monthTotalsOfYear, monthTotals)
-
-			// The status 4 is a report from crawlers that data is unavailable or malformed. By removing them from the API results, we make sure they are displayed as if there is no data.
-			// Fazemos duas checagens no formato do ProcInfo para saber se ele é vazio pois alguns dados diferem, no banco de dados, quando o procinfo é nulo.
-		} else if agencyMonthlyInfo.ProcInfo != nil && agencyMonthlyInfo.ProcInfo.String() != "" && agencyMonthlyInfo.ProcInfo.Status != 4 {
-			monthTotals := models.MonthTotals{Month: agencyMonthlyInfo.Month,
-				BaseRemuneration:   0,
-				OtherRemunerations: 0,
-				CrawlingTimestamp:  agencyMonthlyInfo.CrawlingTimestamp,
-				Error:              &models.ProcError{Stdout: agencyMonthlyInfo.ProcInfo.Stdout, Stderr: agencyMonthlyInfo.ProcInfo.Stderr},
-			}
-			monthTotalsOfYear = append(monthTotalsOfYear, monthTotals)
-		}
-	}
-	sort.Slice(monthTotalsOfYear, func(i, j int) bool {
-		return monthTotalsOfYear[i].Month < monthTotalsOfYear[j].Month
-	})
-	destKey := fmt.Sprintf("%s/datapackage/%s-%d.zip", aID, aID, year)
-	bkp, _ := pgS3Client.Cloud.GetFile(destKey)
-	var pkg *strModels.Package
-	if bkp != nil {
-		pkg = &strModels.Package{
-			AgencyID: &aID,
-			Year:     &year,
-			Package:  *bkp,
-		}
-	}
-
-	agencyTotalsYear := models.AgencyTotalsYear{Year: year, Agency: agency, MonthTotals: monthTotalsOfYear, AgencyFullName: agency.Name, SummaryPackage: pkg}
-	return c.JSON(http.StatusOK, agencyTotalsYear)
-}
-func getBasicInfoOfType(c echo.Context) error {
-	yearOfConsult := time.Now().Year()
-	groupName := c.Param("grupo")
-	var agencies []strModels.Agency
-	var err error
-	var estadual bool
-	var exists bool
-	jurisdicao := map[string]string{
-		"justica-eleitoral":    "Eleitoral",
-		"ministerios-publicos": "Ministério",
-		"justica-estadual":     "Estadual",
-		"justica-do-trabalho":  "Trabalho",
-		"justica-federal":      "Federal",
-		"justica-militar":      "Militar",
-		"justica-superior":     "Superior",
-		"conselhos-de-justica": "Conselho",
-	}
-
-	// Adaptando as URLs do site com o banco de dados
-	// Primeiro consultamos entre as chaves do mapa.
-	if jurisdicao[groupName] != "" {
-		groupName = jurisdicao[groupName]
-	} else {
-		// Caso não encontremos entre as chaves, verificamos entre os valores do mapa.
-		// Isso pois, até a consolidação ser finalizada, o front consulta a api com /Eleitoral, /Trabalho, etc.
-		for _, value := range jurisdicao {
-			if groupName == value {
-				exists = true
-				break
-			}
-		}
-		// Se a jurisdição não existir no mapa, verificamos se trata-se de um estado
-		if !exists {
-			values := map[string]struct{}{"AC": {}, "AL": {}, "AP": {}, "AM": {}, "BA": {}, "CE": {}, "DF": {}, "ES": {}, "GO": {}, "MA": {}, "MT": {}, "MS": {}, "MG": {}, "PA": {}, "PB": {}, "PR": {}, "PE": {}, "PI": {}, "RJ": {}, "RN": {}, "RS": {}, "RO": {}, "RR": {}, "SC": {}, "SP": {}, "SE": {}, "TO": {}}
-			if _, estadual = values[groupName]; estadual {
-				exists = true
-			}
-		}
-		// Se o parâmetro dado não for encontrado de forma alguma, retornamos um NOT FOUND (404)
-		if !exists {
-			return c.JSON(http.StatusNotFound, fmt.Sprintf("Grupo não encontrado: %s.", groupName))
-		}
-	}
-
-	if estadual {
-		agencies, err = pgS3Client.GetStateAgencies(groupName)
-	} else {
-		agencies, err = pgS3Client.GetOPJ(groupName)
-	}
-	if err != nil {
-		// That happens when there is no information on that year.
-		log.Printf("[basic info type] first error getting data for first screen(ano:%d, grupo:%s). Going to try again with last year:%q", yearOfConsult, groupName, err)
-		yearOfConsult = yearOfConsult - 1
-
-		if estadual {
-			agencies, err = pgS3Client.GetStateAgencies(groupName)
-		} else {
-			agencies, err = pgS3Client.GetOPJ(groupName)
-		}
-		if err != nil {
-			log.Printf("[basic info type] error getting data for first screen(ano:%d, grupo:%s):%q", yearOfConsult, groupName, err)
-			return c.JSON(http.StatusBadRequest, fmt.Sprintf("Parâmetros ano=%d ou grupo=%s são inválidos", yearOfConsult, groupName))
-		}
-	}
-	var agenciesBasic []models.AgencyBasic
-	for k := range agencies {
-		agenciesBasic = append(agenciesBasic, models.AgencyBasic{Name: agencies[k].ID, FullName: agencies[k].Name, AgencyCategory: agencies[k].Entity})
-	}
-	state := models.State{Name: c.Param("grupo"), ShortName: "", FlagURL: "", Agency: agenciesBasic}
-	return c.JSON(http.StatusOK, state)
-}
-
-func getSalaryOfAgencyMonthYear(c echo.Context) error {
-	month, err := strconv.Atoi(c.Param("mes"))
-	if err != nil {
-		return c.JSON(http.StatusBadRequest, fmt.Sprintf("Parâmetro mês=%d", month))
-	}
-	year, err := strconv.Atoi(c.Param("ano"))
-	if err != nil {
-		return c.JSON(http.StatusBadRequest, fmt.Sprintf("Parâmetro ano=%d", year))
-	}
-	agencyName := strings.ToLower(c.Param("orgao"))
-	agencyMonthlyInfo, _, err := pgS3Client.GetOMA(month, year, agencyName)
-	if err != nil {
-		log.Printf("[salary agency month year] error getting data for second screen(mes:%d ano:%d, orgao:%s):%q", month, year, agencyName, err)
-		return c.JSON(http.StatusBadRequest, fmt.Sprintf("Parâmetro ano=%d, mês=%d ou nome do orgão=%s são inválidos", year, month, agencyName))
-	}
-	// Fazemos duas checagens no formato do ProcInfo para saber se ele é vazio pois alguns dados diferem, no banco de dados, quando o procinfo é nulo.
-	if agencyMonthlyInfo.ProcInfo != nil && agencyMonthlyInfo.ProcInfo.String() != "" {
-		var newEnv = agencyMonthlyInfo.ProcInfo.Env
-		for _, omittedField := range conf.EnvOmittedFields {
-			for i, field := range newEnv {
-				if strings.Contains(field, omittedField) {
-					newEnv[i] = omittedField + "= ##omitida##"
-					break
-				}
-			}
-		}
-		agencyMonthlyInfo.ProcInfo.Env = newEnv
-		return c.JSON(http.StatusPartialContent, models.ProcInfoResult{
-			ProcInfo:          agencyMonthlyInfo.ProcInfo,
-			CrawlingTimestamp: agencyMonthlyInfo.CrawlingTimestamp,
-		})
-	}
-	return c.JSON(http.StatusOK, models.DataForChartAtAgencyScreen{
-		Members:     agencyMonthlyInfo.Summary.IncomeHistogram,
-		MaxSalary:   agencyMonthlyInfo.Summary.BaseRemuneration.Max,
-		PackageURL:  agencyMonthlyInfo.Package.URL,
-		PackageHash: agencyMonthlyInfo.Package.Hash,
-		PackageSize: agencyMonthlyInfo.Package.Size,
-	})
-}
-
-func getSummaryOfAgency(c echo.Context) error {
-	year, err := strconv.Atoi(c.Param("ano"))
-	if err != nil {
-		return c.JSON(http.StatusBadRequest, fmt.Sprintf("Parâmetro ano=%d inválido", year))
-	}
-	month, err := strconv.Atoi(c.Param("mes"))
-	if err != nil {
-		return c.JSON(http.StatusBadRequest, fmt.Sprintf("Parâmetro mês=%d", month))
-	}
-	agencyName := c.Param("orgao")
-	agencyMonthlyInfo, agency, err := pgS3Client.GetOMA(month, year, agencyName)
-	if err != nil {
-		return c.JSON(http.StatusBadRequest, fmt.Sprintf("Parâmetro ano=%d, mês=%d ou nome do orgão=%s são inválidos", year, month, agencyName))
-	}
-	agencySummary := models.AgencySummary{
-		FullName:   agency.Name,
-		TotalWage:  agencyMonthlyInfo.Summary.BaseRemuneration.Total,
-		MaxWage:    agencyMonthlyInfo.Summary.BaseRemuneration.Max,
-		TotalPerks: agencyMonthlyInfo.Summary.OtherRemunerations.Total,
-		MaxPerk:    agencyMonthlyInfo.Summary.OtherRemunerations.Max,
-		TotalRemuneration: agencyMonthlyInfo.Summary.BaseRemuneration.Total +
-			agencyMonthlyInfo.Summary.OtherRemunerations.Total,
-		TotalMembers: agencyMonthlyInfo.Summary.Count,
-		CrawlingTime: agencyMonthlyInfo.CrawlingTimestamp,
-		HasNext:      time.Date(year, time.Month(month), 1, 0, 0, 0, 0, time.UTC).In(loc).Before(time.Now().AddDate(0, 1, 0)),
-		HasPrevious:  time.Date(year, time.Month(month), 1, 0, 0, 0, 0, time.UTC).In(loc).After(time.Date(2018, 1, 1, 0, 0, 0, 0, time.UTC).In(loc)),
-	}
-	return c.JSON(http.StatusOK, agencySummary)
-}
-
-func generalSummaryHandler(c echo.Context) error {
-	agencies, err := pgS3Client.GetAgenciesCount()
-	if err != nil {
-		return c.JSON(http.StatusInternalServerError, fmt.Sprintf("Erro ao contar orgãos: %q", err))
-	}
-	collections, err := pgS3Client.GetNumberOfMonthsCollected()
-	if err != nil {
-		return c.JSON(http.StatusInternalServerError, fmt.Sprintf("Erro ao contar registros: %q", err))
-	}
-	fmonth, fyear, err := pgS3Client.Db.GetFirstDateWithMonthlyInfo()
-	if err != nil {
-		log.Printf("Error buscando dados - GetFirstDateWithRemunerationRecords: %q", err)
-		return c.JSON(http.StatusInternalServerError, fmt.Sprintf("Erro buscando primeiro registro de remuneração: %q", err))
-	}
-	fdate := time.Date(fyear, time.Month(fmonth), 2, 0, 0, 0, 0, time.UTC).In(loc)
-	lmonth, lyear, err := pgS3Client.GetLastDateWithMonthlyInfo()
-	if err != nil {
-		log.Printf("Error buscando dados - GetLastDateWithRemunerationRecords: %q", err)
-		return c.JSON(http.StatusInternalServerError, fmt.Sprintf("Erro buscando último registro de remuneração: %q", err))
-	}
-	ldate := time.Date(lyear, time.Month(lmonth), 2, 0, 0, 0, 0, time.UTC).In(loc)
-	remuValue, err := pgS3Client.Db.GetGeneralMonthlyInfo()
-	if err != nil {
-		log.Printf("Error buscando dados - GetGeneralRemunerationValue: %q", err)
-		return c.JSON(http.StatusInternalServerError, fmt.Sprintf("Erro buscando valor total de remuneração: %q", err))
-	}
-	return c.JSON(http.StatusOK, models.GeneralTotals{
-		AgencyAmount:             int(agencies),
-		MonthlyTotalsAmount:      int(collections),
-		StartDate:                fdate,
-		EndDate:                  ldate,
-		RemunerationRecordsCount: int(collections),
-		GeneralRemunerationValue: remuValue,
-	})
-}
-
-func getGeneralRemunerationFromYear(c echo.Context) error {
-	year, err := strconv.Atoi(c.Param("ano"))
-	if err != nil {
-		return c.JSON(http.StatusBadRequest, fmt.Sprintf("Parâmetro ano=%d inválido", year))
-	}
-	data, err := pgS3Client.Db.GetGeneralMonthlyInfosFromYear(year)
-	if err != nil {
-		fmt.Println("Error searching for monthly info from year: %w", err)
-		return c.JSON(http.StatusInternalServerError, fmt.Sprintf("Error buscando dados"))
-	}
-	return c.JSON(http.StatusOK, data)
-}
-
-func getAllAgencies(c echo.Context) error {
-	agencies, err := pgS3Client.Db.GetAllAgencies()
-	if err != nil {
-		fmt.Println("Error while listing agencies: %w", err)
-		return c.JSON(http.StatusInternalServerError, "Error while listing agencies")
-	}
-	host := c.Request().Host
-	for i := range agencies {
-		agencies[i].URL = fmt.Sprintf("%s/v1/orgao/%s", host, agencies[i].ID)
-	}
-	return c.JSON(http.StatusOK, agencies)
-}
-
-func getAgencyById(c echo.Context) error {
-	agencyName := c.Param("orgao")
-	agency, err := pgS3Client.Db.GetAgency(agencyName)
-	if err != nil {
-		return c.JSON(http.StatusNotFound, "Agency not found")
-	}
-	host := c.Request().Host
-	agency.URL = fmt.Sprintf("%s/v1/orgao/%s", host, agency.ID)
-	return c.JSON(http.StatusFound, agency)
-}
-
-func getMonthlyInfo(c echo.Context) error {
-	year, err := strconv.Atoi(c.Param("ano"))
-	if err != nil {
-		return c.JSON(http.StatusBadRequest, fmt.Sprintf("Parâmetro ano=%d inválido", year))
-	}
-	agencyName := strings.ToLower(c.Param("orgao"))
-	var monthlyInfo map[string][]strModels.AgencyMonthlyInfo
-	month := c.Param("mes")
-	if month != "" {
-		m, err := strconv.Atoi(month)
-		if err != nil {
-			return c.JSON(http.StatusBadRequest, fmt.Sprintf("Parâmetro mes=%d inválido", m))
-		}
-		oma, _, err := pgS3Client.GetOMA(m, year, agencyName)
-		if err != nil {
-			return c.JSON(http.StatusBadRequest, fmt.Sprintf("Error getting OMA data"))
-		}
-		monthlyInfo = map[string][]strModels.AgencyMonthlyInfo{
-			agencyName: {*oma},
-		}
-	} else {
-		monthlyInfo, err = pgS3Client.Db.GetMonthlyInfo([]strModels.Agency{{ID: agencyName}}, year)
-	}
-	if err != nil {
-		log.Printf("[totals of agency year] error getting data for first screen(ano:%d, estado:%s):%q", year, agencyName, err)
-		return c.JSON(http.StatusBadRequest, fmt.Sprintf("Parâmetro ano=%d ou orgao=%s inválidos", year, agencyName))
-	}
-
-	if len(monthlyInfo[agencyName]) == 0 {
-		return c.NoContent(http.StatusNotFound)
-	}
-
-	// we recreate all struct to controll the serialization better
-
-	type Backup struct {
-		URL  string `json:"url,omitempty"`
-		Hash string `json:"hash,omitempty"`
-		Size int64  `json:"size,omitempty"`
-	}
-	type DataSummary struct {
-		Max     float64 `json:"max,omitempty"`
-		Min     float64 `json:"min,omitempty"`
-		Average float64 `json:"media,omitempty"`
-		Total   float64 `json:"total,omitempty"`
-	}
-	type Summary struct {
-		Count              int         `json:"quantidade,omitempty"`
-		BaseRemuneration   DataSummary `json:"remuneracao_base,omitempty"`
-		OtherRemunerations DataSummary `json:"outras_remuneracoes,omitempty"`
-	}
-	type Summaries struct {
-		MemberActive Summary `json:"membros_ativos,omitempty"`
-	}
-	type Metadata struct {
-		OpenFormat       bool   `json:"formato_aberto"`
-		Access           string `json:"acesso,omitempty"`
-		Extension        string `json:"extensao,omitempty"`
-		StrictlyTabular  bool   `json:"dados_estritamente_tabulares"`
-		ConsistentFormat bool   `json:"manteve_consistencia_no_formato"`
-		HasEnrollment    bool   `json:"tem_matricula"`
-		HasCapacity      bool   `json:"tem_lotacao"`
-		HasPosition      bool   `json:"tem_cargo"`
-		BaseRevenue      string `json:"remuneracao_basica,omitempty"`
-		OtherRecipes     string `json:"outras_receitas,omitempty"`
-		Expenditure      string `json:"despesas,omitempty"`
-	}
-	type Score struct {
-		Score             float64 `json:"indice_transparencia"`
-		CompletenessScore float64 `json:"indice_completude"`
-		EasinessScore     float64 `json:"indice_facilidade"`
-	}
-	type Collect struct {
-		Duration       float64 `json:"duracao_segundos,omitempty"`
-		CrawlerRepo    string  `json:"repositorio_coletor,omitempty"`
-		CrawlerVersion string  `json:"versao_coletor,omitempty"`
-		ParserRepo     string  `json:"repositorio_parser,omitempty"`
-		ParserVersion  string  `json:"versao_parser,omitempty"`
-	}
-
-	type MIError struct {
-		ErrorMessage string `json:"err_msg,omitempty"`
-		Status       int32  `json:"status,omitempty"`
-		Cmd          string `json:"cmd,omitempty"`
-	}
-	type SummaryzedMI struct {
-		AgencyID string     `json:"id_orgao,omitempty"`
-		Month    int        `json:"mes,omitempty"`
-		Year     int        `json:"ano,omitempty"`
-		Summary  *Summaries `json:"sumarios,omitempty"`
-		Package  *Backup    `json:"pacote_de_dados,omitempty"`
-		Meta     *Metadata  `json:"metadados,omitempty`
-		Score    *Score     `json:"indice_transparencia,omitempty`
-		Collect  *Collect   `json:"dados_coleta,omitempty`
-		Error    *MIError   `json:"error,omitempty"`
-	}
-	var summaryzedMI []SummaryzedMI
-	for i := range monthlyInfo {
-		for _, mi := range monthlyInfo[i] {
-			// Fazemos duas checagens no formato do ProcInfo para saber se ele é vazio pois alguns dados diferem, no banco de dados, quando o procinfo é nulo.
-			if mi.ProcInfo == nil || mi.ProcInfo.String() == "" {
-				summaryzedMI = append(
-					summaryzedMI,
-					SummaryzedMI{
-						AgencyID: mi.AgencyID,
-						Error:    nil,
-						Month:    mi.Month,
-						Year:     mi.Year,
-						Package: &Backup{
-							URL:  formatDownloadUrl(mi.Package.URL),
-							Hash: mi.Package.Hash,
-							Size: mi.Package.Size,
-						},
-						Summary: &Summaries{
-							MemberActive: Summary{
-								Count: mi.Summary.Count,
-								BaseRemuneration: DataSummary{
-									Max:     mi.Summary.BaseRemuneration.Max,
-									Min:     mi.Summary.BaseRemuneration.Min,
-									Average: mi.Summary.BaseRemuneration.Average,
-									Total:   mi.Summary.BaseRemuneration.Total,
-								},
-								OtherRemunerations: DataSummary{
-									Max:     mi.Summary.OtherRemunerations.Max,
-									Min:     mi.Summary.OtherRemunerations.Min,
-									Average: mi.Summary.OtherRemunerations.Average,
-									Total:   mi.Summary.OtherRemunerations.Total,
-								},
-							},
-						},
-						Meta: &Metadata{
-							OpenFormat:       mi.Meta.OpenFormat,
-							Access:           mi.Meta.Access,
-							Extension:        mi.Meta.Extension,
-							StrictlyTabular:  mi.Meta.StrictlyTabular,
-							ConsistentFormat: mi.Meta.ConsistentFormat,
-							HasEnrollment:    mi.Meta.HaveEnrollment,
-							HasCapacity:      mi.Meta.ThereIsACapacity,
-							HasPosition:      mi.Meta.HasPosition,
-							BaseRevenue:      mi.Meta.BaseRevenue,
-							OtherRecipes:     mi.Meta.OtherRecipes,
-							Expenditure:      mi.Meta.Expenditure,
-						},
-						Score: &Score{
-							Score:             mi.Score.Score,
-							CompletenessScore: mi.Score.CompletenessScore,
-							EasinessScore:     mi.Score.EasinessScore,
-						},
-						Collect: &Collect{
-							Duration:       mi.Duration,
-							CrawlerRepo:    mi.CrawlerRepo,
-							CrawlerVersion: mi.CrawlerVersion,
-							ParserRepo:     mi.ParserRepo,
-							ParserVersion:  mi.ParserVersion,
-						}})
-				// The status 4 is a report from crawlers that data is unavailable or malformed. By removing them from the API results, we make sure they are displayed as if there is no data.
-			} else if mi.ProcInfo.Status != 4 {
-				summaryzedMI = append(
-					summaryzedMI,
-					SummaryzedMI{
-						AgencyID: mi.AgencyID,
-						Error: &MIError{
-							ErrorMessage: mi.ProcInfo.Stderr,
-							Status:       mi.ProcInfo.Status,
-							Cmd:          mi.ProcInfo.Cmd,
-						},
-						Month:   mi.Month,
-						Year:    mi.Year,
-						Package: nil,
-						Summary: nil,
-						Meta:    nil})
-			}
-		}
-	}
-	return c.JSON(http.StatusOK, summaryzedMI)
-}
-
-func searchByUrl(c echo.Context) error {
-	//Pegando os query params
-	years := c.QueryParam("anos")
-	months := c.QueryParam("meses")
-	agencies := c.QueryParam("orgaos")
-	categories := c.QueryParam("categorias")
-	types := c.QueryParam("tipos")
-	//Criando os filtros a partir dos query params e validando eles
-	filter, err := models.NewFilter(years, months, agencies, categories, types)
-	if err != nil {
-		return c.JSON(http.StatusBadRequest, err.Error())
-	}
-	var category string
-	if filter != nil {
-		category = filter.Category
-	}
-	// Pegando os resultados da pesquisa a partir dos filtros;
-	results, err := postgresDb.Filter(postgresDb.RemunerationQuery(filter), postgresDb.Arguments(filter))
-	if err != nil {
-		log.Printf("Error querying BD (filter or counter):%q", err)
-		return c.JSON(http.StatusInternalServerError, err.Error())
-	}
-	remunerations, numRows, err := getSearchResults(conf.SearchLimit, category, results)
-	if err != nil {
-		log.Printf("Error getting search results: %q", err)
-		return c.JSON(http.StatusInternalServerError, err.Error())
-	}
-
-	response := models.SearchResponse{
-		DownloadAvailable:  numRows > 0 && numRows <= conf.DownloadLimit,
-		NumRowsIfAvailable: numRows,
-		DownloadLimit:      conf.DownloadLimit,
-		SearchLimit:        conf.SearchLimit,
-		Results:            remunerations, // retornando os SearchLimit primeiros elementos.
-	}
-	return c.JSON(http.StatusOK, response)
-}
-
-func downloadByUrl(c echo.Context) error {
-	//Pegando os query params
-	years := c.QueryParam("anos")
-	months := c.QueryParam("meses")
-	agencies := c.QueryParam("orgaos")
-	categories := c.QueryParam("categorias")
-	types := c.QueryParam("tipos")
-
-	//Criando os filtros a partir dos query params e validando eles
-	filter, err := models.NewFilter(years, months, agencies, categories, types)
-	if err != nil {
-		return c.JSON(http.StatusBadRequest, err.Error())
-	}
-
-	results, err := postgresDb.Filter(postgresDb.RemunerationQuery(filter), postgresDb.Arguments(filter))
-	if err != nil {
-		return c.JSON(http.StatusInternalServerError, err.Error())
-	}
-	searchResults, _, err := getSearchResults(conf.DownloadLimit, filter.Category, results)
-	if err != nil {
-		return c.JSON(http.StatusInternalServerError, err.Error())
-	}
-
-	c.Response().Header().Set("Content-Disposition", "attachment; filename=dadosjusbr-remuneracoes.csv")
-	c.Response().Header().Set("Content-Type", c.Response().Header().Get("Content-Type"))
-	err = gocsv.Marshal(searchResults, c.Response().Writer)
-	if err != nil {
-		return c.JSON(http.StatusInternalServerError, fmt.Errorf("erro tentando fazer download do csv: %q", err))
-	}
-	return nil
-}
-
-func getSearchResults(limit int, category string, results []models.SearchDetails) ([]models.SearchResult, int, error) {
-	searchResults := []models.SearchResult{}
-	numRows := 0
-	if len(results) == 0 {
-		return searchResults, numRows, nil
-	} else {
-		// A razão para essa ordenação é que quando o usuário escolhe diversos órgãos
-		// provavelmente ele prefere ver dados de todos eles. Dessa forma, aumentamos
-		// as chances do preview limitado retornar dados de diversos órgãos.
-		sort.SliceStable(results, func(i, j int) bool {
-			return results[i].Ano < results[j].Ano || results[i].Mes < results[j].Mes
-		})
-		searchResults, numRows, err := sess.GetRemunerationsFromS3(limit, conf.DownloadLimit, category, conf.AwsS3Bucket, results)
-		if err != nil {
-			return nil, numRows, fmt.Errorf("failed to get remunerations from s3 %q", err)
-		}
-		return searchResults, numRows, nil
-	}
-}
-
-func formatDownloadUrl(url string) string {
-	return strings.Replace(url, conf.PackageRepoURL, conf.DadosJusURL, -1)
 }
 
 func main() {
@@ -663,23 +117,23 @@ func main() {
 		log.Fatal(err)
 	}
 
-	pgCredentials, err := NewPgCredentials(conf)
+	pgCredentials, err := services.NewPgCredentials(conf.PgUser, conf.PgPassword, conf.PgDatabase, conf.PgHost, conf.PgPort)
 	if err != nil {
 		log.Fatal("Error creating postgres credentials: %v", err)
 	}
 
-	postgresDb, err = NewPostgresDB(*pgCredentials)
+	postgresDb, err = services.NewPostgresDB(*pgCredentials)
 	if err != nil {
 		log.Fatalf("Error connecting to postgres: %v", err)
 	}
 	defer postgresDb.Disconnect()
 
-	postgresDb.conn, err = pgDB.GetConnection()
+	postgresDb.Conn, err = pgDB.GetConnection()
 	if err != nil {
 		log.Fatalf("Error connecting to postgres: %v", err)
 	}
 
-	sess, err = NewAwsSession(conf.AwsRegion)
+	sess, err = services.NewAwsSession(conf.AwsRegion)
 	if err != nil {
 		log.Fatalf("Error creating aws session: %v", err)
 	}
@@ -713,8 +167,8 @@ func main() {
 		if err != nil {
 			log.Fatalf("Error bringin up new relic:%q", err)
 		}
-		postgresDb.newrelic = nr
-		sess.newrelic = nr
+		postgresDb.Newrelic = nr
+		sess.Newrelic = nr
 		uiAPIGroup.Use(nrecho.Middleware(nr))
 		uiAPIGroup.Use(middleware.CORSWithConfig(middleware.CORSConfig{
 			AllowOrigins: []string{"https://dadosjusbr.com", "http://dadosjusbr.com", "https://dadosjusbr.org", "http://dadosjusbr.org", "https://dadosjusbr-site-novo.herokuapp.com", "http://dadosjusbr-site-novo.herokuapp.com"},
@@ -727,34 +181,45 @@ func main() {
 			AllowHeaders: []string{echo.HeaderOrigin, echo.HeaderContentType, echo.HeaderAccept, echo.HeaderContentLength, echo.HeaderAccessControlAllowOrigin},
 		}))
 	}
+	uiApiHandler := handlers.UiApiHandler{
+		Client:           *pgS3Client,
+		PostgresDb:       *postgresDb,
+		Sess:             sess,
+		S3Bucket:         conf.AwsS3Bucket,
+		Loc:              loc,
+		EnvOmittedFields: conf.EnvOmittedFields,
+		SearchLimit:      conf.SearchLimit,
+		DownloadLimit:    conf.DownloadLimit,
+	}
 	// Return a summary of an agency. This information will be used in the head of the agency page.
-	uiAPIGroup.GET("/v1/orgao/resumo/:orgao/:ano/:mes", getSummaryOfAgency)
+	uiAPIGroup.GET("/v1/orgao/resumo/:orgao/:ano/:mes", uiApiHandler.GetSummaryOfAgency)
 	// Return all the salary of a month and year. This will be used in the point chart at the entity page.
-	uiAPIGroup.GET("/v1/orgao/salario/:orgao/:ano/:mes", getSalaryOfAgencyMonthYear)
+	uiAPIGroup.GET("/v1/orgao/salario/:orgao/:ano/:mes", uiApiHandler.GetSalaryOfAgencyMonthYear)
 	// Return the total of salary of every month of a year of a agency. The salary is divided in Wage, Perks and Others. This will be used to plot the bars chart at the state page.
-	uiAPIGroup.GET("/v1/orgao/totais/:orgao/:ano", getTotalsOfAgencyYear)
+	uiAPIGroup.GET("/v1/orgao/totais/:orgao/:ano", uiApiHandler.GetTotalsOfAgencyYear)
 	// Return basic information of a type or state
-	uiAPIGroup.GET("/v1/orgao/:grupo", getBasicInfoOfType)
-	uiAPIGroup.GET("/v1/geral/remuneracao/:ano", getGeneralRemunerationFromYear)
-	uiAPIGroup.GET("/v1/geral/resumo", generalSummaryHandler)
+	uiAPIGroup.GET("/v1/orgao/:grupo", uiApiHandler.GetBasicInfoOfType)
+	uiAPIGroup.GET("/v1/geral/remuneracao/:ano", uiApiHandler.GetGeneralRemunerationFromYear)
+	uiAPIGroup.GET("/v1/geral/resumo", uiApiHandler.GeneralSummaryHandler)
 	// Retorna um conjunto de dados a partir de filtros informados por query params
-	uiAPIGroup.GET("/v2/pesquisar", searchByUrl)
+	uiAPIGroup.GET("/v2/pesquisar", uiApiHandler.SearchByUrl)
 	// Baixa um conjunto de dados a partir de filtros informados por query params
-	uiAPIGroup.GET("/v2/download", downloadByUrl)
+	uiAPIGroup.GET("/v2/download", uiApiHandler.DownloadByUrl)
 
+	apiHandler := handlers.ApiHandler{Client: *pgS3Client, DadosJusURL: conf.DadosJusURL, PackageRepoURL: conf.PackageRepoURL}
 	// Public API configuration
 	apiGroup := e.Group("/v1", middleware.CORSWithConfig(middleware.CORSConfig{
 		AllowOrigins: []string{"*"},
 		AllowHeaders: []string{echo.HeaderOrigin, echo.HeaderContentType, echo.HeaderAccept, echo.HeaderContentLength},
 	}))
 	// Return agency
-	apiGroup.GET("/orgao/:orgao", getAgencyById)
+	apiGroup.GET("/orgao/:orgao", apiHandler.GetAgencyById)
 	// Return all agencies
-	apiGroup.GET("/orgaos", getAllAgencies)
+	apiGroup.GET("/orgaos", apiHandler.GetAllAgencies)
 	// Return MIs by year
-	apiGroup.GET("/dados/:orgao/:ano", getMonthlyInfo)
+	apiGroup.GET("/dados/:orgao/:ano", apiHandler.GetMonthlyInfo)
 	// Return MIs by month
-	apiGroup.GET("/dados/:orgao/:ano/:mes", getMonthlyInfo)
+	apiGroup.GET("/dados/:orgao/:ano/:mes", apiHandler.GetMonthlyInfo)
 
 	s := &http.Server{
 		Addr:         fmt.Sprintf(":%d", conf.Port),
